+++ ESFramework/UIKit/UIViewController+ESAdditions.h
@@ -27,15 +27,6 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (BOOL)isViewVisible;
 
-/**
- * Returns UITabBarController's selected controller or UINavigationController's topViewController,
- * or this controller itself.
- *
- * @note it returns the current "visible" UIViewController, not the "real" controller such as UINavigationController in a UITabBarController.
- *
- */
-- (UIViewController *)currentVisibleViewController;
-
 @end
 
 NS_ASSUME_NONNULL_END