+++ ESFramework/UIKit/UIViewController+ESAdditions.m
@@ -7,61 +7,23 @@
 //
 
 #import "UIViewController+ESAdditions.h"
+#import "NSArray+ESAdditions.h"
 
 @implementation UIViewController (ESAdditions)
 
 - (nullable UIViewController *)previousViewController
 {
-    UINavigationController *navController = self.navigationController;
-    if (navController) {
-        NSArray *controllers = navController.viewControllers;
-        if (0 == controllers.count) {
-            return nil;
-        }
-        NSUInteger index = [controllers indexOfObject:self];
-        if (NSNotFound == index) {
-            return nil;
-        }
-        if (index > 0) {
-            return controllers[index - 1];
-        }
-    }
-    return nil;
+    return [self.navigationController.viewControllers previousObjectToObject:self];
 }
 
 - (nullable UIViewController *)nextViewController
 {
-    UINavigationController *navController = self.navigationController;
-    if (navController) {
-        NSArray *controllers = navController.viewControllers;
-        if (0 == controllers.count) {
-            return nil;
-        }
-        NSUInteger index = [controllers indexOfObject:self];
-        if (NSNotFound == index) {
-            return nil;
-        }
-        if ((index + 1) < controllers.count) {
-            return controllers[index + 1];
-        }
-    }
-
-    return nil;
+    return [self.navigationController.viewControllers nextObjectToObject:self];
 }
 
 - (BOOL)isViewVisible
 {
-    return ([self isViewLoaded] && self.view.window != nil);
-}
-
-- (UIViewController *)currentVisibleViewController
-{
-    if ([self isKindOfClass:[UITabBarController class]]) {
-        return [[(UITabBarController *)self selectedViewController] currentVisibleViewController];
-    } else if ([self isKindOfClass:[UINavigationController class]]) {
-        return [(UINavigationController *)self topViewController];
-    }
-    return self;
+    return self.isViewLoaded && self.view.window != nil;
 }
 
 @end